Make the command type smaller
This makes the command type 16 bytes instead of 24 bytes. This is done
by moving the variable address field into the main enum and making it 32 bits
instead of 64 bits. This commit also changes the usizes in the variables
system to u32s.

I think this is actually nice because it's in the spirit of TeX that the
system be independent of the architecture it's running on. The limitation is
that you can now only have 4GB of memory in count registers....which is fine.
Actually the allocation commands still support more memory.

+++ crates/texlang-stdlib/src/time.rs
@@ -64,8 +64,8 @@ impl Default for Component {
 pub fn get_time<S: HasComponent<Component>>() -> command::VariableFn<S> {
     |_, _, _| -> anyhow::Result<variable::Variable<S>> {
         Ok(variable::Variable::Int(variable::TypedVariable::new(
-            |state: &S, _: usize| -> &i32 { &state.component().minutes_since_midnight },
-            |state: &mut S, _: usize| -> &mut i32 {
+            |state: &S, _: u32| -> &i32 { &state.component().minutes_since_midnight },
+            |state: &mut S, _: u32| -> &mut i32 {
                 &mut state.component_mut().minutes_since_midnight
             },
             0,
@@ -77,8 +77,8 @@ pub fn get_time<S: HasComponent<Component>>() -> command::VariableFn<S> {
 pub fn get_day<S: HasComponent<Component>>() -> command::VariableFn<S> {
     |_, _, _| -> anyhow::Result<variable::Variable<S>> {
         Ok(variable::Variable::Int(variable::TypedVariable::new(
-            |state: &S, _: usize| -> &i32 { &state.component().day },
-            |state: &mut S, _: usize| -> &mut i32 { &mut state.component_mut().day },
+            |state: &S, _: u32| -> &i32 { &state.component().day },
+            |state: &mut S, _: u32| -> &mut i32 { &mut state.component_mut().day },
             0,
         )))
     }
@@ -88,8 +88,8 @@ pub fn get_day<S: HasComponent<Component>>() -> command::VariableFn<S> {
 pub fn get_month<S: HasComponent<Component>>() -> command::VariableFn<S> {
     |_, _, _| -> anyhow::Result<variable::Variable<S>> {
         Ok(variable::Variable::Int(variable::TypedVariable::new(
-            |state: &S, _: usize| -> &i32 { &state.component().month },
-            |state: &mut S, _: usize| -> &mut i32 { &mut state.component_mut().month },
+            |state: &S, _: u32| -> &i32 { &state.component().month },
+            |state: &mut S, _: u32| -> &mut i32 { &mut state.component_mut().month },
             0,
         )))
     }
@@ -99,8 +99,8 @@ pub fn get_month<S: HasComponent<Component>>() -> command::VariableFn<S> {
 pub fn get_year<S: HasComponent<Component>>() -> command::VariableFn<S> {
     |_, _, _| -> anyhow::Result<variable::Variable<S>> {
         Ok(variable::Variable::Int(variable::TypedVariable::new(
-            |state: &S, _: usize| -> &i32 { &state.component().year },
-            |state: &mut S, _: usize| -> &mut i32 { &mut state.component_mut().year },
+            |state: &S, _: u32| -> &i32 { &state.component().year },
+            |state: &mut S, _: u32| -> &mut i32 { &mut state.component_mut().year },
             0,
         )))
     }
