+++ crates/texlang-stdlib/src/tracing.rs
@@ -8,8 +8,8 @@ use texlang_core::variable;
 pub fn get_tracingmacros<S>() -> command::VariableFn<S> {
     |_, _, _| -> anyhow::Result<variable::Variable<S>> {
         Ok(variable::Variable::BaseInt(variable::TypedVariable::new(
-            |base: &BaseState<S>, _: usize| -> &i32 { &base.tracing_macros },
-            |base: &mut BaseState<S>, _: usize| -> &mut i32 { &mut base.tracing_macros },
+            |base: &BaseState<S>, _: u32| -> &i32 { &base.tracing_macros },
+            |base: &mut BaseState<S>, _: u32| -> &mut i32 { &mut base.tracing_macros },
             0,
         )))
     }
